chore: :truck: move tests into central location rather than being split
