+++ sprout/tests/tests_models_tables_and_columns.py
@@ -3,7 +3,7 @@
 from django.test import TestCase
 
 from sprout.app.models import Columns, Tables
-from sprout.app.tests.db_test_utils import (
+from sprout.tests.db_test_utils import (
     create_metadata_table_and_column,
     create_table,
 )
