+++ sprout/tests/tests_views_projects_id_metadata_id_step_name.py
@@ -2,8 +2,8 @@ from django.test import TestCase
 from django.urls import reverse
 
 from sprout.app.models import Tables
-from sprout.app.tests.db_test_utils import create_table
 from sprout.app.views.projects_id_metadata_id.helpers import create_stepper_url
+from sprout.tests.db_test_utils import create_table
 
 
 class MetadataStepNameTests(TestCase):
