+++ sprout/tests/tests_views_table_files.py
@@ -6,7 +6,7 @@ from django.test import TestCase
 from django.urls import reverse
 
 from sprout.app.models import Files
-from sprout.app.tests.db_test_utils import create_table
+from sprout.tests.db_test_utils import create_table
 
 
 class TableFilesTests(TestCase):
